Refactor Settings navigation, add Bug/Feedback to Settings page
